Fix t/tinyish.t when no backend is tested
When skipping all backends the test exit with
fail status.

Adding a dummy 'pass' test could also avoid it.
The SKIP is more about the for loop for backend
but adding it this way make it easier to manage
and see what is tested.

This is fixing an issue noticed on Travis CI when
no backend can be tested.